Avoid potential buffer overflow/missing zero termination of string
Replace a `strncpy(dest, src, strlen(src))` with `strncpy(dest, src,
DEST_SIZE); dest[DEST_SIZE-1]='\0'`. The old code could overflow the
destination if the source was too long, defeating the purpose of
`strncpy`. Further, it failed to put the terminating zero in the
destination, relying on it being properly initialized.